requirements.txt
Django==2.2.5
pymongo==3.12.0
djangorestframework==3.12.4
